langchain/requirements.txt
aiofiles==23.2.1 ; python_full_version >= "3.8.1" and python_version < "3.12"
aiohttp==3.8.5 ; python_full_version >= "3.8.1" and python_version < "3.12"
aiosignal==1.3.1 ; python_full_version >= "3.8.1" and python_version < "3.12"
annotated-types==0.5.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
anyio==3.7.1 ; python_full_version >= "3.8.1" and python_version < "3.12"
async-timeout==4.0.3 ; python_full_version >= "3.8.1" and python_version < "3.12"
attrs==23.1.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
bidict==0.22.1 ; python_full_version >= "3.8.1" and python_version < "3.12"
blinker==1.6.2 ; python_full_version >= "3.8.1" and python_version < "3.12"
brotli==1.1.0 ; platform_python_implementation == "CPython" and python_full_version >= "3.8.1" and python_version < "3.12"
brotlicffi==1.1.0.0 ; platform_python_implementation != "CPython" and python_full_version >= "3.8.1" and python_version < "3.12"
certifi==2023.7.22 ; python_full_version >= "3.8.1" and python_version < "3.12"
cffi==1.15.1 ; platform_python_implementation != "CPython" and python_full_version >= "3.8.1" and python_version < "3.12"
charset-normalizer==3.2.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
click==8.1.7 ; python_full_version >= "3.8.1" and python_version < "3.12"
colorama==0.4.6 ; python_full_version >= "3.8.1" and python_version < "3.12" and platform_system == "Windows"
dataclasses-json==0.6.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
diskcache==5.6.3 ; python_full_version >= "3.8.1" and python_version < "3.12"
duckduckgo-search==3.8.5 ; python_full_version >= "3.8.1" and python_version < "3.12"
exceptiongroup==1.1.3 ; python_full_version >= "3.8.1" and python_version < "3.11"
fastapi==0.103.1 ; python_full_version >= "3.8.1" and python_version < "3.12"
flask-cors==4.0.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
flask-socketio==5.3.6 ; python_full_version >= "3.8.1" and python_version < "3.12"
flask==2.3.3 ; python_full_version >= "3.8.1" and python_version < "3.12"
frozenlist==1.4.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
google-search-results==2.4.2 ; python_full_version >= "3.8.1" and python_version < "3.12"
greenlet==2.0.2 ; python_full_version >= "3.8.1" and python_version < "3.12" and (platform_machine == "win32" or platform_machine == "WIN32" or platform_machine == "AMD64" or platform_machine == "amd64" or platform_machine == "x86_64" or platform_machine == "ppc64le" or platform_machine == "aarch64")
h11==0.14.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
h2==4.1.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
hpack==4.0.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
httpcore==0.18.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
httpx[brotli,http2,socks]==0.25.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
hyperframe==6.0.1 ; python_full_version >= "3.8.1" and python_version < "3.12"
idna==3.4 ; python_full_version >= "3.8.1" and python_version < "3.12"
importlib-metadata==6.8.0 ; python_full_version >= "3.8.1" and python_version < "3.10"
itsdangerous==2.1.2 ; python_full_version >= "3.8.1" and python_version < "3.12"
jinja2==3.1.2 ; python_full_version >= "3.8.1" and python_version < "3.12"
jsonpatch==1.33 ; python_full_version >= "3.8.1" and python_version < "3.12"
jsonpointer==2.4 ; python_full_version >= "3.8.1" and python_version < "3.12"
langchain==0.0.300 ; python_full_version >= "3.8.1" and python_version < "3.12"
langsmith==0.0.40 ; python_full_version >= "3.8.1" and python_version < "3.12"
llama-cpp-python==0.2.6 ; python_full_version >= "3.8.1" and python_version < "3.12"
lxml==4.9.3 ; python_full_version >= "3.8.1" and python_version < "3.12"
markupsafe==2.1.3 ; python_full_version >= "3.8.1" and python_version < "3.12"
marshmallow==3.20.1 ; python_full_version >= "3.8.1" and python_version < "3.12"
multidict==6.0.4 ; python_full_version >= "3.8.1" and python_version < "3.12"
mypy-extensions==1.0.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
numexpr==2.8.6 ; python_full_version >= "3.8.1" and python_version < "3.12"
numpy==1.24.4 ; python_full_version >= "3.8.1" and python_version < "3.12"
openai==0.28.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
packaging==23.1 ; python_full_version >= "3.8.1" and python_version < "3.12"
pycparser==2.21 ; python_full_version >= "3.8.1" and platform_python_implementation != "CPython" and python_version < "3.12"
pydantic-core==2.6.3 ; python_full_version >= "3.8.1" and python_version < "3.12"
pydantic==2.3.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
python-dotenv==1.0.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
python-engineio==4.7.1 ; python_full_version >= "3.8.1" and python_version < "3.12"
python-socketio==5.9.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
pyyaml==6.0.1 ; python_full_version >= "3.8.1" and python_version < "3.12"
requests==2.31.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
simple-websocket==0.10.1 ; python_full_version >= "3.8.1" and python_version < "3.12"
sniffio==1.3.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
socksio==1.0.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
sqlalchemy==2.0.21 ; python_full_version >= "3.8.1" and python_version < "3.12"
starlette==0.27.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
tenacity==8.2.3 ; python_full_version >= "3.8.1" and python_version < "3.12"
tqdm==4.66.1 ; python_full_version >= "3.8.1" and python_version < "3.12"
typing-extensions==4.8.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
typing-inspect==0.9.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
urllib3==2.0.5 ; python_full_version >= "3.8.1" and python_version < "3.12"
uvicorn==0.23.2 ; python_full_version >= "3.8.1" and python_version < "3.12"
werkzeug==2.3.7 ; python_full_version >= "3.8.1" and python_version < "3.12"
wsproto==1.2.0 ; python_full_version >= "3.8.1" and python_version < "3.12"
yarl==1.9.2 ; python_full_version >= "3.8.1" and python_version < "3.12"
zipp==3.17.0 ; python_full_version >= "3.8.1" and python_version < "3.10"
